The last commit introduced a bug where if the ASE filter and LBO scan at the same time, the ASE scan was very slow (once per second).
The problem was, that the ASE Tab and the LBO Autoscan both measure the wavelength. Since the wavelength was not scanned, only the GetWavelength method was executed in the while-Loop over and over, which was too much for the WLM.

IMPLEMENTED FIX: Added a small (0.1s) sleep-timer after the wavelength measurement in the LBO Autoscan method.